% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-prep.R
\name{util.agg.value}
\alias{util.agg.value}
\title{generalized aggregation function}
\usage{
util.agg.value(x, agg.func = max)
}
\arguments{
\item{x}{input series/array}

\item{agg.func}{the aggregation function}
}
\value{
the aggregated value
}
\description{
generalized aggregation function, handy when use together with summarise_all
}
\examples{
util.agg.value(c(rep(1,4), rep(2, 10)), agg.func = mean)
util.agg.value(c(rep('a',4), rep('b', 10)), agg.func = util.get.mode)
library(dplyr)
mtcars \%>\% summarise_all(function(i) util.agg.value(i, util.get.mode))
}
