% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-prep.R
\name{util.cap.outlier}
\alias{util.cap.outlier}
\title{cap outliers within a certain range}
\usage{
util.cap.outlier(x, verbose = FALSE)
}
\arguments{
\item{x}{input numeric array}

\item{verbose}{boolean, TRUE -> show cap limits}
}
\value{
capped numeric array
}
\description{
cap outliers within the range of Q1/Q3 +/- 1.5 * IQR
}
\examples{
util.cap.outlier(c(rnorm(10), 100), verbose = TRUE)
}
