% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.corr.compute}
\alias{util.corr.compute}
\title{util.corr.compute}
\usage{
util.corr.compute(data, fnames)
}
\arguments{
\item{data}{input data.frame}

\item{fnames}{variables to perform correlation computation}
}
\value{
data.frame
}
\description{
compute pairwise correlation with p-value
}
\examples{
util.corr.compute(mtcars, c('am','gear','mpg'))
}
