% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.corr.plot}
\alias{util.corr.plot}
\title{plot pairwise correlation}
\usage{
util.corr.plot(out, title = "Pairwise Correlation Plot")
}
\arguments{
\item{out}{data.frame output from `util.corr.compute`}

\item{title}{title of the chart}
}
\value{
plot
}
\description{
plot pairwise correlation, used together with `util.corr.compute`
}
\examples{
out <- util.corr.compute(mtcars, colnames(mtcars))
util.corr.plot(out)
}
