% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{util.file.summ}
\alias{util.file.summ}
\title{tabulate the file summary from `ls` command}
\usage{
util.file.summ(dir)
}
\arguments{
\item{dir}{directory to be summarized}
}
\value{
data.frame
}
\description{
output the file summary from bash command `ls`
}
\examples{
util.file.summ('./')

}
