% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-prep.R
\name{util.get.mode}
\alias{util.get.mode}
\title{compute mode number of a numeric array}
\usage{
util.get.mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{input array}

\item{na.rm}{if remove na}
}
\value{
the mode number
}
\description{
return the mode number
}
\examples{
util.get.mode(c(rep(1,4), rep(2, 10)))
util.get.mode(c(rep('a',4), rep('b', 10)))
}
