% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{util.init.dir}
\alias{util.init.dir}
\title{util.init.dir}
\usage{
util.init.dir(path)
}
\description{
create a directory/file if it doesn't exist
}
\examples{
util.init.dir('./tmp1/')
util.init.dir('./tmp2/text.txt')
unlink('tmp1', recursive = T)
unlink('tmp2', recursive = T)
}
