% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{util.init.logger}
\alias{util.init.logger}
\title{util.init.logger}
\usage{
util.init.logger(
  logger = "",
  level = "INFO",
  log.file = NULL,
  reset.log = TRUE
)
}
\arguments{
\item{logger}{name of logger}

\item{level}{The logging level of the root logger.}

\item{log.file}{external path of log file}

\item{reset.log}{boolean. TRUE -> reset logger}
}
\description{
initiator for logger functions
}
\examples{
util.init.logger()
logging::loginfo('this is an info')
logging::logwarn('this is a warning')
logging::logerror('this is an error')
}
