% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{util.lh2.fill}
\alias{util.lh2.fill}
\title{default color code}
\usage{
util.lh2.fill(code = NA)
}
\arguments{
\item{code}{either int or string. \cr
1,'red \cr
2,'orange' \cr
3,'yellow' \cr
4,'lightgreen' \cr
5,'green' \cr
anything else, default color \cr}
}
\value{
color code
}
\description{
default color code
}
\examples{
util.lh2.fill()
}
