% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.lh2.theme}
\alias{util.lh2.theme}
\title{util.lh2.theme}
\usage{
util.lh2.theme(angle = 45, legend.position = "right", font.size = 12)
}
\arguments{
\item{angle}{int, rotation of x.axis.text}

\item{legend.position}{char, {'right','left','top','bottom','none'}}

\item{font.size}{int font size of text}
}
\value{
a ggplot2 theme
}
\description{
default theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) + 
  geom_point() + 
  util.lh2.theme()
}
