% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.plot.stack.bar}
\alias{util.plot.stack.bar}
\title{util.plot.stack.bar}
\usage{
util.plot.stack.bar(df, x, fill, label.threshold = 0.1, ...)
}
\arguments{
\item{df}{data.frame, input data}

\item{x}{char, variable to plot for statistics}

\item{fill}{char, variable to partition by color}

\item{label.threshold}{float, only label percentage greater than threshold}

\item{...}{extra param for function `geom_bar`}
}
\description{
plot stack bar with proprtion percentage for two categorical varaibles
}
\examples{
mtcars['am'] <- as.factor(mtcars[['am']])
mtcars['gear'] <- as.factor(mtcars[['gear']])
util.plot.stack.bar(mtcars, 'gear', 'am')
}
