% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.plots.share.axis}
\alias{util.plots.share.axis}
\title{combine plots and share axis}
\usage{
util.plots.share.axis(
  ...,
  shared.axis = "y",
  plots.theme = NULL,
  plots.coord = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  nrow = NULL,
  ncol = NULL,
  legend.position = "bottom",
  return.plots = FALSE
)
}
\arguments{
\item{...}{ggplot objects}

\item{shared.axis}{{'x', 'y'},  defaults to 'y'}

\item{plots.theme}{ggplot theme, defaults to \code{util.lh2.theme}}

\item{plots.coord}{ggplot coord, defaults to \code{coord_flip}}

\item{title}{str, plot global title}

\item{subtitle}{str, plot global subtitle}

\item{caption}{str, plot global caption}

\item{nrow}{int, number of rows of the layout}

\item{ncol}{int, number of ncols of the layout}

\item{legend.position}{{'right','left','bottom','top'}}

\item{return.plots}{if TRUE -> return list of plots only, defaults to FALSE}
}
\value{
ggplot
}
\description{
modify plots by hiding x/y axes if shared x/y axis is used, plots combination
is done by \code{patchwork} package. \cr
plotlist can also be output by setting \code{return.plots = TRUE}.
}
\examples{
library(ggplot2)
library(dplyr)
library(patchwork)

p1 <- mtcars \%>\%
  mutate(gear = factor(gear), am = factor(am)) \%>\%
  util.plot.stack.bar('gear','am') +
  labs(title = 'p1')
p2 <- mtcars \%>\%
  mutate(am = factor(gear)) \%>\%
  group_by(am) \%>\%
  summarise(n = n()) \%>\%
  ggplot(aes(x = am, y = n)) +
  geom_bar(stat = 'identity', fill = util.lh2.fill('orange'))+
  labs(title = 'p2')
  
util.plots.share.axis(p1, p2, 
  title = 'Shared Axis Plot')
}
