% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.profile.cat}
\alias{util.profile.cat}
\title{util.profile.cat}
\usage{
util.profile.cat(data, x, segment = "segment", alpha = 0.8)
}
\arguments{
\item{data}{input data.frame}

\item{x}{char, categorical variable name}

\item{segment}{variable name of the segment}

\item{alpha}{float density plot transparency}
}
\value{
list(data, plot)
}
\description{
profile the numeric variable by segment, used for segment profiling
}
\examples{
res <- util.profile.cat(mtcars, x = 'am', segment = 'gear')
res$data
res$plot
}
