% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.profile.fnames}
\alias{util.profile.fnames}
\title{util.profile.fnames}
\usage{
util.profile.fnames(data, fnames, segment, alpha = 0.8, scale = 1)
}
\arguments{
\item{data}{input data.frame}

\item{fnames}{variables to be profiled}

\item{segment}{variable name of the segment}

\item{alpha}{float density plot transparency}

\item{scale}{float scale of density plot}
}
\value{
list(data, plot)
}
\description{
profile both numeric and categorical varaibles
}
\examples{
mtcars['am'] <- factor(mtcars[['am']])
res <- util.profile.fnames(mtcars, c('am','wt'), segment = 'gear')
library(ggpubr)
ggarrange(plotlist = res$plot, ncol = 2 , nrow = 1, common.legend = TRUE)
}
