% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.profile.num}
\alias{util.profile.num}
\title{density plot for single numeric variable}
\usage{
util.profile.num(data, x, segment = "segment", alpha = 0.8, scale = 1)
}
\arguments{
\item{data}{input data.frame}

\item{x}{numeric variable for profiling}

\item{segment}{variable name of the segment}

\item{alpha}{float density plot transparency}

\item{scale}{float scale of density plot}
}
\value{
list(data.frame, plot)
}
\description{
profile the numeric variable colored by segment, used for segment profiling
}
\examples{
res <- util.profile.num(mtcars, x = 'wt', segment = 'am')
res$data
res$plot

}
