% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.segment.summarize}
\alias{util.segment.summarize}
\title{summarize segments with custom aggregation function}
\usage{
util.segment.summarize(
  data,
  segment,
  summ.func = NULL,
  highlight.tol = 0.2,
  kable = TRUE,
  ...
)
}
\arguments{
\item{data}{output from `util.profile.fnames`}

\item{segment}{char, segment variable name}

\item{summ.func}{segment summarize function `e.g function(i) summarise(i, n = n())`}

\item{highlight.tol}{tolerance for highlighting}

\item{kable}{boolean, TRUE -> return kable table}

\item{...}{extra params from `util.vis.kable`}
}
\description{
summarize segments in tables
}
\examples{
util.segment.summarize(mtcars, segment = 'gear')
library(dplyr)
summ.func <- function(i){
  summarise(i, n = n(), 
               `am-0` = mean(am == 0),
               `am-1` = mean(am == 1))
}
util.segment.summarize(mtcars, 'gear', summ.func)
}
