% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-table.R
\name{util.text.wrap}
\alias{util.text.wrap}
\title{break long text}
\usage{
util.text.wrap(x, wrap.len = 50, br = "<br>")
}
\arguments{
\item{x}{input text}

\item{wrap.len}{the max of string length before adding line break}

\item{br}{line break separator}
}
\value{
character
}
\description{
add line break in text
}
\examples{
x <- paste(rep('a',20), collapse = '')
util.text.wrap(x, wrap.len = 5)
}
