% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{util.theme.no.axis}
\alias{util.theme.no.axis}
\title{Add-on theme without x/y axis}
\usage{
util.theme.no.axis(axis = "x")
}
\arguments{
\item{axis}{char, {'x','y'}}
}
\value{
ggplot theme
}
\description{
quick function to remove the x/y axis
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) + 
  geom_point() + 
  util.lh2.theme() +
  util.theme.no.axis('x')
}
