% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-table.R
\name{util.vis.datatable}
\alias{util.vis.datatable}
\title{util.vis.datatable}
\usage{
util.vis.datatable(
  df,
  type = c(),
  dom = "frtip",
  options = NULL,
  scrollY = 400,
  rownames = F,
  filter = "none",
  background.color = "#008B8B",
  pretty.header = F,
  ...
)
}
\arguments{
\item{df}{input data.frame}

\item{type}{character, {'buttons','scroll-y'}}

\item{dom}{character, a combined string to indicate table elements}

\item{options}{list, extra options from `DT` js package}

\item{scrollY}{int, number of pixels to the scrollable window}

\item{rownames}{boolean, TRUE -> show rowname of table}

\item{filter}{character, {'none','top','bottom'}, position of filter}

\item{background.color}{character, header color}

\item{pretty.header}{boolean, TRUE -> enable colored header}

\item{...}{extra params from `datatable` function}
}
\value{
rendered html table
}
\description{
visualize table using kable package
}
\examples{
util.vis.datatable(head(mtcars), type = 'buttons')
util.vis.datatable(head(mtcars), type = 'scroll-y')
}
