% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-table.R
\name{util.vis.kable}
\alias{util.vis.kable}
\title{util.vis.kable}
\usage{
util.vis.kable(
  df,
  convert_pct = F,
  accuracy = 1,
  full_width = F,
  escape = F,
  caption = NULL,
  align = NULL,
  digits = 1,
  position = "center",
  pretty.header = T,
  ...
)
}
\arguments{
\item{df}{input data.frame}
}
\value{
rendered html table
}
\description{
visualize table using kable package
}
\examples{
util.vis.kable(head(mtcars))
}
