% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{util.init.logger}
\alias{util.init.logger}
\title{util.init.logger}
\usage{
util.init.logger(
  logger = "",
  level = "INFO",
  log.file = NULL,
  reset.log = TRUE
)
}
\description{
initiator for logger functions
}
\examples{
util.init.logger()
}
