% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-table.R
\name{util.vis.datatable}
\alias{util.vis.datatable}
\title{util.vis.datatable}
\usage{
util.vis.datatable(
  df,
  type = c(),
  dom = "frtip",
  options = NULL,
  scrollY = 400,
  rownames = F,
  filter = "none",
  background.color = "#008B8B",
  pretty.header = F,
  ...
)
}
\arguments{
\item{df}{input data.frame}
}
\value{
rendered html table
}
\description{
visualize table using kable package
}
\examples{
util.vis.datatable(head(mtcars), type = 'buttons')
util.vis.datatable(head(mtcars), type = 'scroll-y')
}
