% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-table.R
\name{util.vis.kable}
\alias{util.vis.kable}
\title{util.vis.kable}
\usage{
util.vis.kable(
  df,
  convert_pct = F,
  accuracy = 1,
  full_width = F,
  escape = F,
  caption = NULL,
  align = NULL,
  align.left = 1L,
  digits = 1,
  position = "center",
  pretty.header = T,
  ...
)
}
\arguments{
\item{df}{input data.frame}

\item{convert_pct}{boolean, TRUE -> convert number to percent if possible}

\item{accuracy}{decimal for percentage, e.g. 1/0.1/0.01}

\item{full_width}{boolean, TRUE -> full width table}

\item{escape}{boolean, TRUE escape html tags}

\item{caption}{character, caption of the table}

\item{align}{character, a combined string of text alignment. e.g. 'lcrr' for 4 columns}

\item{align.left}{int, number of columns left aligned.}

\item{digits}{int, number of decimal digits to keep}

\item{position}{{'left','right','float_left','center'}}

\item{pretty.header}{boolean, TRUE -> colored header}

\item{...}{extra params from kable_styling function}
}
\value{
rendered html table
}
\description{
visualize table using kable package
}
\examples{
util.vis.kable(head(mtcars))
}
